requirements.txt
Flask
opencv-python
numpy
keras
requests
gunicorn
setuptools==75.8.0
wheel==0.45.1
tensorflow
python-dotenv